fix: document correction and refactored
