+++ src/pages/privileges/outlets/linixUseCase/types.ts
@@ -1,3 +1,5 @@
+import { IAssignmentFormEntry } from "../users/types/forms.types";
+
 interface OpcionesCsPorCasoDeUso {
   k_Opcion: string;
 }
@@ -31,7 +33,7 @@ interface UseCase {
   opcionesCsPorCasoDeUso?: OpcionesCsPorCasoDeUso[];
   opcionesPortalWebPorCasoDeUso?: OpcionesPortalWebPorCasoDeUso[];
   reportesCsPorCasoDeUso?: ReportesCsPorCasoDeUso[];
-  reportesWebPorCasoDeUso?: ReportesWebPorCasoDeUso[];
+  reportesWebPorCasoDeUso?: IAssignmentFormEntry[];
   tiposDeDocumentoPorCasoDeUso?: TiposDeDocumentoPorCasoDeUso[];
 }
